+++ changelog.md
@@ -9,9 +9,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [1.6.5] - 2026-05-26
+### ⚡ Updates
 
-## [1.6.4] - 2026-05-13
+- Toned down logging on the `ORMConnectionProvider` to avoid noisy logs on every connection acquisition.
+- Dropped "No ORM application found" logs when ORM is disabled.
+
+## [1.6.5] - 2026-05-26
 
 ### ⭐ Added